Inital commit
This commit is massive but this is the result of massive spike and stabilize
and the history should not do anything good for anyone.

This repo started of as a way to learn Kotlin and is heavily inspired by GitVersion
adding the assumption that a RC tag will be added at develop when using GitFlow makes
it possible to implement this in a much easier way and old code that was using a .version
file and even trying to backtrack the release branches has been removed.